# Conflicts:
#	examples/Cartesian Mesh/CMakeLists.txt
#	examples/Cartesian Mesh/main.cpp
#	examples/ECLIPSE Mesh/CMakeLists.txt
#	examples/ECLIPSE Mesh/main.cpp
#	examples/GMSH Unstructured Mesh/CMakeLists.txt
#	examples/GMSH Unstructured Mesh/main.cpp
#	examples/INRIA Unstructured Mesh/main.cpp

+++ PAMELA/source/Parallel/ParallelEnsemble.hpp
@@ -68,7 +68,7 @@ namespace PAMELA
 		collection_iterator end_ghost() { return m_data.end(); }
 
 		//Operators
-		T &operator[](int i) { return m_data[i]; }
+		T &operator[](size_t i) { return m_data[i]; }
 
 
 		//Push back T
